// shared/types.ts
export interface Player {
  id: string;
  name: string;
  position: number; // 0 to 39
  balance: number;
  isBankrupt: boolean;
  inJail: boolean;
  jailTurns: number;
  getOutOfJailFreeCards?: number;
  skipTurns?: number;
  powerCards?: string[]; // Array to hold power cards like 'BECOME_A_DON'
  avatar: string; // Color identifier or token icon
  loan?: {
    principal: number;
    interestRate: number;
    totalRepayment: number;
    remainingAmount: number;
    remainingTurns: number;
    deductionPerTurn: number;
  };
}